Use SVGs with PNG fallbacks for toolbar icons
Simplify the SVG source as it is now being served to the
client. The InkScape metadata contained nothing of use.

With the simplified SVG we can now create the color variants
using a simple text replace in the SVG.

These SVGs can then be used to create the colored PNGs directly
which removes the dependency on ImageMagick.

Fixes #1137.