Remove z-index rule to .leaflet-pane
`leaflet-pane` is applied to not only each layer element (which bear their own
z-index rules) but also to the parent container for them. This lead to
ordering issues where custom interfaces over top a map now need agressive
z-index rules and consequently broke examples.

This commit drops the z-index applied to this general classname. It's carry
over from Leaflet/Leaflet#3591 which appears to be a little half hearted.