Run make to create built release before NPM publish
At the moment each project depending on mapbox.js has to create the built release files on their own. With that we bring the built releases directly to these projects by putting them inside the package before publish, so they are available for everyone directly when the package is installed via npm.

The git repository instead is still clean and will not contain them.